requirements.txt
# Core dependencies
google-generativeai>=0.3.0
chromadb>=0.4.0
networkx>=3.0
python-dotenv>=1.0.0

# Document processing
pypdf>=3.17.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web interface
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0

# Utilities
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0

# Agent System (Optional - for agent-based architecture)
redis>=5.0.0  # For distributed message bus (optional, falls back to in-memory)

# Optional: For advanced features
# spacy>=3.7.0  # For better NER
# sentence-transformers>=2.2.0  # For alternative embeddings
# neo4j>=5.14.0  # For Neo4j graph backend
